Use INI for application settings on all platforms
Previously the application settings were stored in different formats,
depending on the operating system. But this is a bit cumbersome,
especially the Windows Registry is a pain. So now the INI file format
is used on all platforms.

In addition, the location of the INI file can be specified with the
environment variable LIBREPCB_CONFIG_DIR (useful for functional tests).